MITAB: Add CPLErrorReset() to the start of TABMAPFile::Open()
Fix autotest2 OgrMitabTest2.testMapInfoFileWithoutIndex test by
resetting the last error count at the beginning of TABMAPFile::Open().

This is needed because the test input file has charset of "WindowsLatin1",
which gets converted to encoding "CP1252". However, in one of the conversion
routines, it adds a warning:

  Warning 1: Recode from CP1252 to UTF-8 not supported, treated as ISO8859-1 to UTF-8.

The warning by itself is not a problem, but it calls CPLError(), which records
the last error. This last error gets carried over to TABMAPFile::Open(), which
by itself is fine. But at the end of function, it checks whether there was any
last error reported. And this fails because of the above charset encoding
warning.


git-svn-id: https://svn.osgeo.org/gdal/trunk@39725 f0d54148-0727-0410-94bb-9a71ac55c965