Remove hacky method to run scripts as imported
The *need* for this hack is unclear. Docstring states it's so the interpreter
doesn't need to guess it's own name, but the `_as_external_script` variant is
alive and well and being used in other parts of the codebase, so it works fine.

This hack seems to play weirdly with pytest (maybe something to do with the output
capturing?) and it seemed best to just remove it.