Disable a LC_CTYPE test in test_gdal2tiles
This test is tricky in that it tests some non-nominal behaviour of gdal2tiles
regarding encoding settings. The failure here must be due to using an external
process now instead of the current python interpreter.

The issue now is that the 'if not os.path.isfile(input_file)' check
at line 1211 of gdal2tiles choke on non-ASCII input, whereas previously
it was a unicode string at that point, and the check at line 1252 would
emit the WARNING.

In other words, this probably never worked to begin with, but the old test
behaviour masked that.