OGRBuildPolygonFromLines: avoid generating effectively duplicate points
Floating-point rounding error can mean that the beginning and end
coordinates of a ring do not exactly match, for example
123456.036528677,5234567.04091659 vs 123456.036528678,5234567.04091659.
Repeating the first point at the end of the ring may lead to a tiny
self-intersecting segment, making the geometry technically invalid.

A similar issue arose when combining line segments into the ring.

The supplied tolerance is now used to determine whether points should be
treated as equal. If this proves problematic, a hard-coded tolerance
expression that is proportional to the values being compared (such as
poLine->getX(0) * 1e-14) could be used instead.