gdal2tiles: Refactor and fix multiprocessing completion handling
The mulitiprocessing.Pool has an imap_unordered method that can be used
to propagate completion events back to the main thread, which is cleaner
and less error-prone than handling it manually with a queue and running
the progress bar in an additional process.

This fixes a bug when combining the `--exclude` option with
`--processes`: the skipped fully transparent tiles were not properly
writing a completion event to the queue, so the progress bar did not
reflect these skipped tiles, and the process hung waiting to receive the
expected number of results from the queue. This also avoids similar
hangs if a worker process were to throw an exception.

By batching together tiles with `chunksize`, there is a small
performance improvement due to better cache utilization when processing
adjacent tiles in the same process.