GTiff: make -co NUM_THREADS produce reproducible output
Currently the blocks being compressed are flushed to disk in the order
the worker threads complete. Which may be different from the order in
which they were submitted. This is not desirabe and this commit makes
sure that the blocks are flushed to disk in the order they were requested
to be written.