nitf/nitfimage.c: skip bad UDID data (fixes #1578) (#1586)
When reading UDID and IXSHD metadata, the two TRE record sets are
concatenated in memory.  If the UDID data is not a valid TRE set, the
IXSHD was not being parsed.  Now, if the UDID data does not have valid
TRE length as the first entry, it is skipped with a warning and the
IXSHD data is still parsed.

This adds two test files that are modified from the fake_nsif.ntf test
file.  The valid_udid.ntf file has a CSDIDA record added to a UDID
block; the invalid_udid.ntf file has a UDID block that does not contain
a TRE.

Fixes #1578.