Block cache: fix corruption on multithreaded write on datasets (fixes #2011)
When writing several GDAL datasets in parallel from several threads, a
thread can force a dirty block of another thread/dataset to be written. By
doing so, it will remove it from the list of cached blocks, so if the thread
that explictly manipulates this dataset reacquire the block, it will see an
old version of the block. The effect of this can be particularly seen with
GeoTIFF files with multi band pixel leaving organization, where acquiring
a block of a band causes corresponding blocks of other bands to be cached.

The main fix done here is in GDALRasterBlock::Internalize() to avoid evicting
a dirty block of another dataset when there are other candidates. This considerably
reduces the likelihood of the bug.

Another fix is done in GDALDataset::EnterReadWrite() to wait for the completion
of the flushing of pending dirty blocks of the dataset before reading new blocks.

I believe there could still be problems if the block cache size is too
small regarding the number of threads.