s57objectclasses.csv: add wtwdis and unlocd for distance marks from IENC (#2123)
I found the values by inspecting the object with opencpn as chart-viewer

The s57 data are the official inland waterway maps from the german waterway
administration. They are provided free of charge.

e.g. Mittellandkanal https://en.wikipedia.org/wiki/Mittelland_Canal
  https://www.elwis.de/DE/dynamisch/IENC/download_commit.php?file=WW41_2018-06-22_0_318

or an overview at:
  https://www.elwis.de/DE/Service/Inland-ENC-der-WSV/Inland-ENC-der-WSV-node.html

or an example without the hassel of extracting
  http://gpl.coulmann.de/tmp/1W7ML310.000

The chart holds river kilometers with not physically installed distance
marks. I like to use this for personal interest for locking up the
nearest river kilometer if I have an AIS gps position of a vessel.