Enable Bash completions and control installation
bash-completion expects completions to be installed in a given directory
and named for program being completed.

Detect the directories as suggested in the FAQ:
https://github.com/scop/bash-completion/blob/master/README.md#faq

Control installation with a new configuration option --with-bash-completion.
If it is not specified nothing will be installed.
If a path is passed completions will be installed in that directory.

Install the GDAL completion script as gdalinfo and create symlinks to it
for the other programs.