Update dted.rst
The DTED driver reports that it only supports CreateCopy() via the metadata information. This is backed up by the fact that up until GDAL 3.1 you were unable to use gdalwarp directly to output a file in DTED format. GDAL 3.1 introduced the ability to output in formats that supported CreateCopy() and you can now use gdalwarp with the DTED format.