OSM: Don't filter out explicit attributes (fixes #2603)
If the user explicitly lists an attribute as desired in osmconf.ini,
it must not be filtered out early.
Note that the attribute list is configured per object type, but the
early filter is global. Listing a desired attribute for one type may
make it appear also for other types, unless it is explicitly in the
other types' ignore list.