PG and PostgisRaster drivers: set the application name in PostgreSQL connections (#2658)
If the application name is not set in the connection string
(and PGAPPNAME environment variable is not set),
then set the application name to the current GDAL version.
PostgreSQL connections can take a long time and they do show
up in the database session list. Being able to identify database
sessions makes diagnostics easier.