VSICurlHandle: Fixed error found with the pcidsk_fuzzer.
The pcidsk_fuzzer exposed a bug in VSICurlHandle::Read(). The values were: iterOffset=280, nOffsetToDownload=0, osRegion.size()=272. The error is probably a side effect of the PCIDSK format trying to read something pass the end-of-file by disregarding the return values of Seek() and Read(). But this change can do no harm and it removes a vunerability. I also ran the gtiff_fuzzer for a day and there were no errors.