ENVI, Ehdr, etc: lower memory req. for BIP interleave, and improve efficiency
Raw datasets with BIP interleaving used to allocate for each band a
buffer of raster_width * data_type_size * nbands, which resulted in a
quadratic RAM allocation in number of bands. This could prevent opening
of large files and resulted in suboptimal speed at reading/writing.

This fix doesn't change the C++ ABI, at the expense of an internal
trick.

Proper fix for #3213